rename core/apis files to follow naming conventions
