--- lib/core/apis/EcoleDirecte/convertersExporter.dart
@@ -1,8 +1,0 @@
-export 'package:ynotes/core/apis/EcoleDirecte/converters/cloud.dart';
-export 'package:ynotes/core/apis/EcoleDirecte/converters/disciplines.dart';
-export 'package:ynotes/core/apis/EcoleDirecte/converters/documents.dart';
-export 'package:ynotes/core/apis/EcoleDirecte/converters/homework.dart';
-export 'package:ynotes/core/apis/EcoleDirecte/converters/lessons.dart';
-export 'package:ynotes/core/apis/EcoleDirecte/converters/mail.dart';
-export 'package:ynotes/core/apis/EcoleDirecte/converters/account.dart';
-export 'package:ynotes/core/apis/EcoleDirecte/converters/schoolLife.dart';
