--- lib/core/apis/Pronote/convertersExporter.dart
@@ -1,4 +1,0 @@
-export 'package:ynotes/core/apis/Pronote/converters/disciplines.dart';
-export 'package:ynotes/core/apis/Pronote/converters/homework.dart';
-export 'package:ynotes/core/apis/Pronote/converters/documents.dart';
-export 'package:ynotes/core/apis/Pronote/converters/lessons.dart';
