+++ lib/core/apis/ecole_directe/converters/homework.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:ynotes/core/apis/EcoleDirecte/convertersExporter.dart';
+import 'package:ynotes/core/apis/ecole_directe/converters_exporter.dart';
 import 'package:ynotes/core/logic/modelsExporter.dart';
 
 class EcoleDirecteHomeworkConverter {
