+++ lib/core/apis/ecole_directe/converters/lessons.dart
