+++ lib/ui/screens/login/widgets/login_web_view.dart
@@ -6,7 +6,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
-import 'package:ynotes/core/apis/Pronote/PronoteCas.dart';
+import 'package:ynotes/core/apis/pronote/Pronote_cas.dart';
 import 'package:ynotes/core/apis/utils.dart';
 import 'package:ynotes/globals.dart';
 import 'package:ynotes/main.dart';
@@ -110,12 +110,12 @@ class _LoginWebViewState extends State<LoginWebView> {
                 child: Column(
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: [
-                    if (!kIsWeb  &&Platform.isLinux)
+                    if (!kIsWeb && Platform.isLinux)
                       Text(
                         "La connexion par ENT n'est pas encore supportée sur Linux...",
                         style: TextStyle(fontFamily: "Asap", color: Colors.red),
                       ),
-                    if (!kIsWeb  && !Platform.isLinux)
+                    if (!kIsWeb && !Platform.isLinux)
                       Text(
                         "Patientez... nous vous connectons à l'ENT",
                         style: TextStyle(fontFamily: "Asap"),
